+++ hf2p_server/source/networking/network_utilities.cpp
@@ -1,6 +1,6 @@
 #include "network_utilities.h"
 
-void network_get_build_identifiers(int* executable_type, int* executable_version, int* compatible_version)
+void network_get_build_identifiers(long* executable_type, long* executable_version, long* compatible_version)
 {
 	*executable_type = 5;
 	*executable_version = 11;
