+++ hf2p_server/source/networking/transport/transport_security.h
@@ -4,7 +4,7 @@
 
 struct s_transport_secure_address
 {
-	uint32_t  data1;
+	uint32_t data1;
 	uint16_t data2;
 	uint16_t data3;
 	byte data4[8];
@@ -23,4 +23,4 @@ struct s_transport_secure_key
 	char data[16];
 };
 
-int transport_secure_address_get(GUID* transport_secure_address);
+long transport_secure_address_get(GUID* transport_secure_address);
